Convert :cake-convert and :cake-clean to command types
Lift the two cake handlers into CakeConvertCommand and CakeCleanCommand (discrete
commands, names preserved) taking IConsolePrompts via the constructor. The .cake
syntax-rewriting helpers (GetCakeFiles/GetCakeConvertedContent/GetCakePackages)
stay on Program as internal residual — GetCakeConvertedContent/GetCakePackages are
exercised directly by the Cake conversion tests — until the #392 collapse PR.
Replaces both DelegateCommand adapters. Tests: 30 pass.
