+++ src/Fallout.Cli/Program.cs
@@ -66,8 +66,11 @@ public partial class Program
 
         // Legacy handlers still living on Program, adapted until they are extracted into command
         // types. Each conversion deletes one line here plus its Program.X.cs partial.
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("cake-convert", CakeConvert));
-        services.AddSingleton<IFalloutCommand>(new DelegateCommand("cake-clean", CakeClean));
+        services.AddSingleton<IFalloutCommand, CakeConvertCommand>();
+        services.AddSingleton<IFalloutCommand, CakeCleanCommand>();
+
+        // Legacy handlers still living on Program, adapted until they are extracted into command
+        // types. Each conversion deletes one line here plus its Program.X.cs partial.
         services.AddSingleton<IFalloutCommand>(new DelegateCommand("GetNextDirectory", GetNextDirectory));
         services.AddSingleton<IFalloutCommand>(new DelegateCommand("PopDirectory", PopDirectory));
         services.AddSingleton<IFalloutCommand>(new DelegateCommand("PushWithCurrentRootDirectory", PushWithCurrentRootDirectory));
